pyproject.toml
[project]
name = "mcp-server-inspector"
version = "0.1.0"
description = "MCP Server Inspector Tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "mcp>=1.12.3",
    "anthropic>=0.50.0",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.5.0",
]

[project.scripts]
inspector = "main:main"
